requirements.txt
# AURORA-X Dependencies
# Core
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0
pydantic>=2.0

# Machine Learning & Deep Learning
torch>=2.1.0
scikit-learn>=1.3.0
onnx>=1.14.0
onnxruntime>=1.16.0

# Streaming & Messaging
aiokafka>=0.9.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Database & Caching
sqlalchemy>=2.0
aiosqlite>=0.19.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# Time-series
influxdb-client>=1.37.0

# MLflow
mlflow>=2.8.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Compression
lz4>=4.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
rich>=13.0
click>=8.1.0
